+++ test/worker_tasks/one_arg_tasks.dart
@@ -1,5 +1,11 @@
+import 'dart:developer';
+
 Future<void> voidOneArgTask(_) async {}
 
 Future<String> valueTask(String value) async => value;
 
 Future<int> incrementTask(int value) async => value + 1;
+
+Future<UserTag> unsupportedArgumentTask(_) async {
+  return UserTag("");
+}